== exp calculations & items == *Update of the level up calculations. -Updated to use uint64 for player calculations. -Updated part of caclulations with Hercules as reference. -Thanks to @HerculesWS & @Reilaen for the base.
*Some item database updates
-Fixed Revolver_Buffalo_Bandit_Card script.
-Fixed Marlin trade restrictions.

+++ src/common/mmo.h
@@ -576,7 +576,7 @@ struct mmo_charstatus {
 	uint32 mother;
 	uint32 child;
 
-	unsigned int base_exp,job_exp;
+	uint64 base_exp,job_exp;
 	int zeny;
 
 	short class_;
